requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
distro==1.9.0
fastapi==0.115.6
uvicorn[standard]==0.34.0
genson==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
jsonpath-ng==1.7.0
openai==1.99.3
ply==3.11
pydantic==2.11.7
python-dotenv==1.0.0
pydantic_core==2.33.2
PyYAML==6.0.2
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
watchdog==5.0.3
